Inline connection parsing into connector's FS
